+++ cmaes_integration/run_cmaes.py
@@ -2,19 +2,23 @@
 Runs cma-es on `run_simulation.py` as a fitness function.
 Creates output.csv and updates it continuously with the best individual from each generation.
 Whether to show the simulation or save as video, number of generations, sigma can be passed as
-command line arguments. Example: `python3 run_cmaes.py headless 50 2` runs cma-es for 50 generations
-in headless mode with a sigma of 2. Replacing "headless" with "screen" makes the simulation 
-output to the screen, and replacing it with "video" saves each simulation as a video in `./videos`. 
-"both" shows on screen and saves a video.
+command line arguments. Example: `python3 run_cmaes.py --gens 50 --sigma 2 --mode h` 
+runs cma-es for 50 generations
+in headless mode with a sigma of 2. Replacing "--mode h" with "--mode s" makes the simulation 
+output to the screen, and replacing it with "--mode v" saves each simulation 
+as a video in `./videos`. 
+"-mode b" shows on screen and saves a video.
 
-Author: Thomas Breimer
+Authors: Thomas Breimer, James Gaskell
 February 4th, 2025
 """
 
-import sys, csv, argparse
+import csv
+import argparse
 from cmaes import CMA
 import numpy as np
 import run_simulation as sim
+
 
 def run_cma_es(mode, gens, sigma_val):
     """
@@ -38,8 +42,9 @@ def run_cma_es(mode, gens, sigma_val):
     csv_header = ['generation', 'best_fitness']
 
     for i in range(sim.NUM_ACTUATORS):
-        csv_header = csv_header + ['frequency' + str(i), 'amplitude' + str(i),
-                             'phase_offset' + str(i)]
+        csv_header = csv_header + [
+            'frequency' + str(i), 'amplitude' + str(i), 'phase_offset' + str(i)
+        ]
 
     with open("output.csv", "w", newline="") as file:
         writer = csv.writer(file)
@@ -47,15 +52,17 @@ def run_cma_es(mode, gens, sigma_val):
 
     # Init CMA
 
-    optimizer = CMA(mean=np.array([sim.AVG_FREQ, sim.AVG_AMP, sim.AVG_PHASE_OFFSET]
-                                   * sim.NUM_ACTUATORS), sigma=sigma_val)
+    optimizer = CMA(mean=np.array(
+        [sim.AVG_FREQ, sim.AVG_AMP, sim.AVG_PHASE_OFFSET] * sim.NUM_ACTUATORS),
+                    sigma=sigma_val)
 
     for generation in range(gens):
         solutions = []
 
         for indv_num in range(optimizer.population_size):
             x = optimizer.ask()
-            fitness = sim.run(sim.NUM_ITERS, x, mode, str(generation) + "_" + str(indv_num))
+            fitness = sim.run(sim.NUM_ITERS, x, mode,
+                              str(generation) + "_" + str(indv_num))
             solutions.append((x, fitness))
 
         optimizer.tell(solutions)
@@ -73,19 +80,17 @@ def run_cma_es(mode, gens, sigma_val):
 if __name__ == "__main__":
     parser = argparse.ArgumentParser(description='RL')
     parser.add_argument(
-        '--mode', #headless, screen, video, both h, s, v, b
+        '--mode',  #headless, screen, video, both h, s, v, b
         help='mode for output. h-headless , s-screen, v-video, b-both',
         default="h")
-    parser.add_argument(
-        '--gens',
-        type=int,
-        help='number of generations to run',
-        default=100)
-    parser.add_argument(
-        '--sigma',
-        type=int,
-        default=2,
-        help='sigma value for cma-es')
+    parser.add_argument('--gens',
+                        type=int,
+                        help='number of generations to run',
+                        default=100)
+    parser.add_argument('--sigma',
+                        type=int,
+                        default=2,
+                        help='sigma value for cma-es')
     args = parser.parse_args()
 
     run_cma_es(args.mode, args.gens, args.sigma)